backend_service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pika==1.3.2
pytest==7.4.3
